WIP refactoring into multiple files
